Create BasicPoWAlgorithm.cpp
Creating a Proof-of-Work (PoW) consensus algorithm involves complex cryptographic operations and handling of blockchain data. This is typically done in a high-level language with access to cryptographic libraries and resources.

However, here's a very basic scaffold of a PoW algorithm-like structure in C++. This example does not include any actual cryptocurrency functionality, but it gives an idea of the structure a PoW algorithm might have.

This code defines Transaction, Block, and Blockchain classes to represent transactions, blocks, and the blockchain. The Block class includes a mineBlock method that represents the PoW algorithm. In the main function, a Blockchain object is created, a block with a transaction is mined and added to the blockchain.