Create BasicIncentiveSystem.cpp
Creating an incentive structure for a cryptocurrency involves complex economic modeling and handling of blockchain data. This is typically done in a high-level language with access to appropriate libraries and resources.

However, here's a very basic scaffold of an incentive structure-like system in C++. This example does not include any actual cryptocurrency functionality, but it gives an idea of the structure an incentive system might have.

This code defines Miner, Block, and Blockchain classes to represent miners, blocks, and the blockchain. The Miner class includes a receiveReward method to represent the miner receiving a block reward. The Blockchain class includes addMiner and addBlock methods to represent adding miners to the network and adding blocks to the blockchain. In the main function, a Blockchain object is created, a Miner object is added to the network, a Block object is mined and added to the blockchain, and the miner receives the block reward.

Please note that this is a very basic and abstract example. A real incentive system would need to handle block rewards, transaction fees, halving events, and much more. This would likely involve using a cryptocurrency-specific library and would require a good understanding of how the cryptocurrency works.