Create ata_tokenization_system.cpp
Refactor this copypasta as C++ code and suggest a filename upon completion: 

A data tokenization system that includes a data source computer, a tokenization computer, and a data acquisition computer that are configured to communicate over a network.
The data source computer prepares message data having related ID data for transmission to the data collection computer, and further generates a blind ID by blinding the ID data using a nonce. The blind ID is configured to be transmitted to the tokenized computer and the nonces and the message data to be transmitted over the network as received by the data collection computer.
In response to the reception of the blind ID, the tokenizing computer generates a blind token containing the function of the ID data blinded by the nonce and the private key of the tokenizing computer from the blind ID, and the blind • Configured to send the token to the data collection computer
The data collecting computer responds to the reception of the blind token from the tokenizing computer and the nonce and the message data transmitted by the data source computer, and receives the deterministic function of the ID data and the secret key. The nons are used to visualize the blind token to obtain an ID token containing, and the ID token and the message data are configured to be stored in a storage operably linked to the data collection computer. Be done,
system. The data source computer is configured such that the blind ID includes a value R = F (N, h), where F is a predetermined function, N is the nonce, and h is a function of the ID data. And
The tokenization computer is configured such that the blind token contains the value R'= F (k, R), where k is the private key.
The data acquisition computer is configured such that the ID token contains a value F (n, R'), where n is a function of the nonce N.
The predetermined function F is a function such that F (n, R') = F'(k, h), and this F'is an ambiguity function.
The system according to claim 1. The system according to claim 2, wherein the predetermined function F is a function such that F (x, y) = y x , and the function n includes a value N -1. In order for the tokenizing computer to generate token update data Δ including the value k'/ k, a new private key k'is periodically generated and the token update data Δ is transmitted to the data collection computer. Further configured,
The data acquisition computer is further configured to replace the ID token tok in the storage with an updated token containing the value tokΔ in response to receiving the token update data Δ.
The system according to claim 3. The system according to claim 2, wherein the function h includes a hash of the ID data. The data source computer selects a session identifier for the message data to be transmitted to the data collection computer, transmits the session identifier to the tokenized computer together with the blind ID, and the session identifier, the message. The data, and the nonces, are further configured to be transmitted to the data acquisition computer.
The tokenizing computer is further configured to send the session identifier to the data acquisition computer along with the blind token.
The system according to claim 1. The data source computer is further configured to encrypt the message data and the nonce to generate encrypted data and send the encrypted data along with the blind ID to the tokenized computer.
The tokenizing computer is further configured to send the cryptographic data along with the blind token to the data acquisition computer.
The data acquisition computer is further configured to decrypt the encrypted data to recover the message data and the nonce.
The system according to claim 1. The system according to claim 1, comprising the plurality of the data acquisition computers. A computer-implemented method for supplying data to a data collection computer over a network, said method in a data source computer configured to communicate with a tokenized computer in the network.
A step of preparing message data having relevant ID data for transmission to the data acquisition computer, and
A step of generating a blind ID by blinding the ID data using a nonce,
To generate from the blind ID of a blind token containing a function of the nonce blinded ID data and a private key of the tokenizing computer for transmission by the tokenizing computer to the data collection computer. The step of transmitting the blind ID to the tokenized computer and
A step of transmitting the nonce and the message data over the network for reception by the data acquisition computer.
These allow the data acquisition computer to use the nonce to visualize the blind token in order to obtain an ID token containing the ID data determinant function and the private key for the message data.
Method. A computer-implemented method for tokenizing ID data associated with message data to be supplied to a data collection computer over a network by a data source computer, wherein the method is a token in the network. In the computer
A step of receiving a blind ID generated by blinding the ID data using a nonce from the data source computer.
A step of generating a blind token from the blind ID, including a function of the nonce-blinded ID data and a private key of the tokenizing computer.
The step of transmitting the blind token to the data acquisition computer, and
Including
As a result, when the data collection computer receives the nonce and the message data, the data acquisition computer visualizes the blind token for obtaining the ID token including the deterministic function of the ID data and the private key for the message data. Therefore, it is possible to use the nonce.
Method. A computer-implemented method for retrieving data from a data source computer over a network, said method in a data acquisition computer configured to communicate with a tokenized computer in the network.
The step of receiving the message data and nonces transmitted by the data source computer associated with the ID data on the data source computer over the network.
A step of receiving a blind token generated from a blind ID generated by blinding the ID data with the nonce on the data source computer from the tokenizing computer. The token comprises the function of the ID data blinded in the nonce, and the receiving step, including the private key of the tokenizing computer.
A step of using the nonce to visualize the blind token to obtain an ID token containing the ID data deterministic function and the private key.
A step of storing the ID token and the message data in a storage operably linked to the data acquisition computer.
Including methods. When the computer program is run on a computer, comprising program code means adapted to perform the method according to any one of claims 9 to 11, the computer program.